requirements.txt
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-csp==4.0
envtoml==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
psycopg2-binary==2.9.10
pwned-passwords-django==5.2.0
pycparser==2.23
python-json-logger==4.0.0
sniffio==1.3.1
sqlparse==0.5.3
toml==0.10.2
cryptography==42.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-axes==6.5.1
drf-spectacular==0.28.0
factory-boy==3.3.3
pytest==8.3.4
pytest-django==4.10.0
pytest-cov==6.0.0
python-json-logger==4.0.0
pyotp==2.9.0
